[FIX] UAASDEV-551, High CPU load on database server
External process call our Odoo instance with sessions that they discard within
half a minute after which they fetch /web/login without a database argument.
This causes Odoo to fetch the list of available databases with a postgresql
connection that is not reused. These 64 connections or so per minute cause a
much higher load on the database than is necessary.

For our purposes, we can simply cache the list of available databases. If the
list of available databases is manipulated from within Odoo we can clear the
cached list. Only if a database is created directly in Postgresql, Odoo needs
to be restarted to become aware of it.