[FIX] UAASDEV-711, Multiple connector heartbeat threads give concurrency errors
Because we start various threads during the loading of the Odoo registry,
we have to rely on the fact that the registry is loaded only once in the
supervisor process and then forked off to every spawned worker.
However, when multiple nodes are spawned of simultaneously, setting up the
qualified count function will raise a specific error, breaking the loading
of the registry in the supervisor thread. The registry will then be loaded
in each of the worker threads separately, leading to multiple hearbeat
threads and AMPQ listeners.