[RFR] Keep track of seen dependencies when installing a module.
In a hugely dense dependency tree, installing a module with 143 dependencies
can lead to 2.5 million module checks when insisting on checking
each dependency of each dependency recursively, halting the module
installation process for an hour or more. Keeping track of seen modules
increases the performance so that checking the dependencies finishes
instantly.